refactored code (add controllers)
