+++ controllers/index.js
@@ -1,0 +1,3 @@
+const products = require('./contacts')
+
+module.exports = { products }
