+++ middlewares/index.js
@@ -1,0 +1,3 @@
+const controllerWrapper = require('./controllerWrapper')
+
+module.exports = { controllerWrapper }
